created a new project in android studio
Chose min sdk version to be 21. Based this decision off the fact that the michigan hackers app had to have a lot of workarounds to support versions less than 21. Also, 21 allows for lock screen notifications, which seems to be something we could end up implementing. According to https://developer.android.com/about/dashboards/, 21+ will support about 90% of people that visit the google play store, and I believe this number is higher for those in the US.