working on sign-in logic
Added FirebaseAuthActivity for activities that must ensure that a user is signed in.
Added signout button in mainactivity for testing.
Moved sign in with google/facebook to ServiceLoginHandler class so the same logic can be used for LoginActivity.
Currently incorrectly signing out, causing app to crash when trying to sign out 2x in a row.