Style fixes
Fix size of most everything after realizing that the sizes in figma corresponded to dp and sp.
Increased size of textinputlayout hint and error text.
Fixed opacity of the white colors in the login pages.
Currently, stroke weight and left margins on edit text hints are each the same across all login pages. This is not the case for the figma design, and I haven't yet figured out if this is a mistake.
The register with google and register with facebook buttons on the registration page also don't quite match teh figma design yet.