update profile pic and display name working
The way to edit the display name is by clicking pencil icon next to edittext. It saves or displays error snackbar if you click outside of edit text or click enter in keyboard. If you try to swipe away snackbar displayed above keyboard, it will recognize it as clicking away from edit text and close the keyboard.