Profile Activity styling
Added style for padding left/right/top for screens because I believe all screens will have same padding except for login screens. Login screens padding was seet in LinearLayoutLoginPages, but I moved it to OUtermostLayoutPaddingLoginPages to make it consistent with OutermostLayoutPadding. Pulling the padding out of the linear layout style for the login pages might be overkill.
Styled everything in the profile activity above the divider, except still don't have "Add Friend" and "Block" buttons.
Made a "Header" style to be used for the headers across all the screens e.g. Friends, Players, Remaining, Nearby Players, etc.
Made a reusable layout for the profile image and add image FAB.
Added ripple effect to back button. One small problem is that the ripple effect on the left side of the back button is cut off because the back button has a negative left margin in order to position is correctly.
Added me.grantland.autofittextview dependency because I couldn't get android's native autofit to work for edittexts. I believe it might only work for textviews. This was necessary to fit longer display names on the screen. On my phone, only about 9 characters could fit at 36sp, but our character limit is 20. Didn't want to just reduce character limit to 9 because it could vary by screen size.
Also made a reusable layout for the divider because it is also used in the "invite players" screen.